# docs/swagger.yaml
definitions:
...
  models.CategoryResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.CreatePostRequest:
    properties:
      body:
        maxLength: 8142
        type: string
      categoryIds:
        items:
          type: integer
        type: array
      thumbnail:
        type: string
      title:
        maxLength: 256
        type: string
    required:
    - body
    - title
    type: object
  models.PostResponse:
    properties:
      body:
        maxLength: 8142
        type: string
      categories:
        items:
          $ref: '#/definitions/models.CategoryResponse'
        type: array
      createdAt:
        type: string
      id:
        type: integer
      thumbnail:
        type: string
      title:
        maxLength: 256
        type: string
      updatedAt:
        type: string
    required:
    - body
    - title
    type: object
  models.SignInRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 64
        minLength: 5
        type: string
    required:
    - email
    type: object
  models.SignInResponse:
    properties:
      token:
        type: string
    type: object
  models.SignUpRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        maxLength: 64
        minLength: 5
        type: string
    required:
    - email
    - name
    type: object
  models.SignUpResponse:
    properties:
      token:
        type: string
    type: object
...
  models.UpdatePostRequest:
    properties:
      body:
        maxLength: 8142
        type: string
      categoryIds:
        items:
          type: integer
        type: array
      thumbnail:
        type: string
      title:
        maxLength: 256
        type: string
    required:
    - body
    - title
    type: object
  models.UpdateUserRequest:
    properties:
      name:
        type: string
      password:
        type: string
      photo:
        type: string
    required:
    - name
    - password
    type: object
  models.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      photo:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact:
    email: tubagus.sflh@gmail.com
    name: API Support
    url: http://github.com/bagus2x
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /post:
    post:
      description: Creating a new post.
      parameters:
      - description: the body for creating post
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CreatePostRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PostResponse'
      summary: create a new post.
      tags:
      - Post
  /post/{postId}:
    delete:
      description: deleting a post.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post id
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: delete single post.
      tags:
      - Post
    get:
      description: finding a post by id.
      parameters:
      - description: Post id
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostResponse'
      summary: get a category by id.
      tags:
      - Post
    put:
      description: updating a post.
      parameters:
      - description: the body for updating category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePostRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post id
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostResponse'
      summary: update single post.
      tags:
      - Post
...
  /user:
    delete:
      description: Delete authenticated user.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete a user.
      tags:
      - User
    put:
      description: Update authenticated user.
      parameters:
      - description: the body for updating user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SignUpResponse'
      summary: Update a user.
      tags:
      - User
  /user/{userId}:
    get:
      description: Finding user by id.
      parameters:
      - description: Project id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
      summary: get a user by id.
      tags:
      - User
  /user/signin:
    post:
      description: Gaining access to the protected API. Put the token from response
        in the Authorization header to access protected API.
      parameters:
      - description: the body to sign in
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SignInResponse'
      summary: log in with registered email.
      tags:
      - User
  /user/signup:
    post:
      description: Creating a new account to authenticate a user.
      parameters:
      - description: the body for sign up
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SignUpResponse'
      summary: create a new account.
      tags:
      - User
...
swagger: "2.0"